requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.3.0
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
fake-useragent==0.1.11
idna==2.8
importlib-metadata==1.4.0
isodate==0.6.0
jsonpickle==1.2
lxml==4.4.2
mimesis==3.3.0
more-itertools==8.1.0
packaging==20.0
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytest==5.3.4
pytz==2019.3
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
wcwidth==0.1.8
wstools-py3==0.54.4
zeep==3.4.0
zipp==2.0.0
